+++ lib/components/sprites/home_field.dart
@@ -1,14 +1,13 @@
 import 'dart:ui';
 
 import 'package:EileMitWeile/components/player.dart';
-import 'package:EileMitWeile/components/tokens.dart';
+import 'package:EileMitWeile/components/token.dart';
 import 'package:flame/components.dart';
-import '../eilemitweile_game.dart';
+import '../../eilemitweile_game.dart';
 
-class HomeField extends PositionComponent with HasGameRef<EilemitweileGame>  {
+class HomeField extends PositionComponent with HasGameRef<EilemitweileGame> {
   late Player player;
   List<Token> tokens = [];
-
 
   @override
   bool get debugMode => false;
@@ -21,12 +20,16 @@ class HomeField extends PositionComponent with HasGameRef<EilemitweileGame>  {
   @override
   void render(Canvas canvas) {
     if (player.color == 1)
-      homefield_red.render(canvas, position: Vector2(0, 0), anchor: Anchor.topLeft);
+      homefield_red.render(canvas,
+          position: Vector2(0, 0), anchor: Anchor.topLeft);
     else if (player.color == 2)
-      homefield_blue.render(canvas, position: Vector2(0, 0), anchor: Anchor.topLeft);
+      homefield_blue.render(canvas,
+          position: Vector2(0, 0), anchor: Anchor.topLeft);
     else if (player.color == 3)
-      homefield_green.render(canvas, position: Vector2(0, 0), anchor: Anchor.topLeft);
+      homefield_green.render(canvas,
+          position: Vector2(0, 0), anchor: Anchor.topLeft);
     else
-      homefield_yellow.render(canvas, position: Vector2(0, 0), anchor: Anchor.topLeft);
+      homefield_yellow.render(canvas,
+          position: Vector2(0, 0), anchor: Anchor.topLeft);
   }
 }