+++ lib/components/text_components/dicenumber.dart
@@ -9,7 +9,8 @@ class ScoreText extends TextComponent with HasGameRef<EilemitweileGame> {
     this.text_content = "",
   }) : _textPaint = TextPaint(
             textDirection: TextDirection.rtl,
-            style: TextStyle(fontSize: 80, color: Colors.black));
+            style: TextStyle(
+                fontSize: 80, color: Colors.black, fontFamily: 'Komika'));
 
   late final TextPaint _textPaint;