+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.0.4
+- Fix: some refactoring
+- Feature: Show triple 6 message
+- Fix: comicify graphics
+- Add fonts
+
 ## v0.0.3
 
 - Feature: Throw again with a 6